呼叫中心sql整改

+++ guiyu-parent/guiyu-botstence-parent/guiyu-botstence-dal/src/main/java/ai/guiji/botsentence/dao/BotAvailableTemplateMapper.java
@@ -4,14 +4,13 @@ import ai.guiji.botsentence.dao.entity.BotAvailableTemplate;
 import ai.guiji.botsentence.dao.entity.BotAvailableTemplateExample;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
-import org.springframework.web.bind.annotation.RequestParam;
 
 public interface BotAvailableTemplateMapper {
     int countByExample(BotAvailableTemplateExample example);
 
     int deleteByExample(BotAvailableTemplateExample example);
 
-    int deleteByPrimaryKey(Long id);
+    int deleteByPrimaryKey(Integer id);
 
     int insert(BotAvailableTemplate record);
 
@@ -19,7 +18,7 @@ public interface BotAvailableTemplateMapper {
 
     List<BotAvailableTemplate> selectByExample(BotAvailableTemplateExample example);
 
-    BotAvailableTemplate selectByPrimaryKey(Long id);
+    BotAvailableTemplate selectByPrimaryKey(Integer id);
 
     int updateByExampleSelective(@Param("record") BotAvailableTemplate record, @Param("example") BotAvailableTemplateExample example);
 
@@ -28,9 +27,9 @@ public interface BotAvailableTemplateMapper {
     int updateByPrimaryKeySelective(BotAvailableTemplate record);
 
     int updateByPrimaryKey(BotAvailableTemplate record);
-    
+
     List<BotAvailableTemplate> getUserAvailableTemplate(Long userId);
-    
+
     void addUserAvailableTemplate(@Param("userId")Long userId,@Param("array")String[] availableId);
     void addUserAvailableTemplateAuto(@Param("userId")Long userId,@Param("availableId")String availableId);
